requirements.txt
numpy>=1.23
pillow
matplotlib>=3.5
pandas>=1.5.0
openpyxl>=3.0.0
